refactor: Phase 1 — decompose tide/models/deployment monoliths
Split tide.py, models.py, and deployment.py into focused modules with
re-export shims for backward compatibility. Relocate CBC/Splunk init mixins
into Engines/modules/systems/ with legacy path shims.
